Extract Roborazzi utils into `:roborazzi` module
And cleanup unnecessary transitive dependencies.

Fixes #1161
