+++ core/database/build.gradle.kts
@@ -34,5 +34,6 @@ dependencies {
 
     implementation(libs.kotlinx.datetime)
 
-    androidTestImplementation(projects.core.testing)
+    androidTestImplementation(libs.androidx.test.core)
+    androidTestImplementation(libs.kotlinx.coroutines.test)
 }
