+++ core/designsystem/build.gradle.kts
@@ -46,9 +46,7 @@ dependencies {
     testImplementation(libs.accompanist.testharness)
     testImplementation(libs.hilt.android.testing)
     testImplementation(libs.robolectric)
-    testImplementation(libs.roborazzi)
-    testImplementation(projects.core.testing)
+    testImplementation(projects.roborazzi)
 
     androidTestImplementation(libs.androidx.compose.ui.test)
-    androidTestImplementation(projects.core.testing)
 }
