+++ core/testing/build.gradle.kts
@@ -15,7 +15,6 @@
  */
 plugins {
     alias(libs.plugins.nowinandroid.android.library)
-    alias(libs.plugins.nowinandroid.android.library.compose)
     alias(libs.plugins.nowinandroid.android.hilt)
 }
 
@@ -24,23 +23,15 @@ android {
 }
 
 dependencies {
-    api(kotlin("test"))
-    api(libs.androidx.compose.ui.test)
-    api(libs.roborazzi)
+    api(libs.kotlinx.coroutines.test)
+
     api(projects.core.analytics)
+    api(projects.core.common)
     api(projects.core.data)
     api(projects.core.model)
     api(projects.core.notifications)
 
-    debugApi(libs.androidx.compose.ui.testManifest)
-
-    implementation(libs.accompanist.testharness)
-    implementation(libs.androidx.activity.compose)
     implementation(libs.androidx.test.rules)
     implementation(libs.hilt.android.testing)
-    implementation(libs.kotlinx.coroutines.test)
     implementation(libs.kotlinx.datetime)
-    implementation(libs.robolectric.shadows)
-    implementation(projects.core.common)
-    implementation(projects.core.designsystem)
 }
