+++ feature/bookmarks/build.gradle.kts
@@ -29,5 +29,6 @@ dependencies {
 
     testImplementation(projects.core.testing)
 
+    androidTestImplementation(libs.androidx.compose.ui.test)
     androidTestImplementation(projects.core.testing)
 }
