+++ feature/foryou/build.gradle.kts
@@ -32,6 +32,9 @@ dependencies {
     testImplementation(libs.hilt.android.testing)
     testImplementation(libs.robolectric)
     testImplementation(projects.core.testing)
+    testImplementation(projects.roborazzi)
 
+
+    androidTestImplementation(libs.androidx.compose.ui.test)
     androidTestImplementation(projects.core.testing)
 }
