+++ feature/foryou/src/test/kotlin/com/google/samples/apps/nowinandroid/feature/foryou/ForYouScreenScreenshotTests.kt
@@ -21,15 +21,15 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import com.google.samples.apps.nowinandroid.core.designsystem.component.NiaBackground
 import com.google.samples.apps.nowinandroid.core.designsystem.theme.NiaTheme
-import com.google.samples.apps.nowinandroid.core.testing.util.DefaultTestDevices
-import com.google.samples.apps.nowinandroid.core.testing.util.captureForDevice
-import com.google.samples.apps.nowinandroid.core.testing.util.captureMultiDevice
 import com.google.samples.apps.nowinandroid.core.ui.NewsFeedUiState
 import com.google.samples.apps.nowinandroid.core.ui.NewsFeedUiState.Success
 import com.google.samples.apps.nowinandroid.core.ui.UserNewsResourcePreviewParameterProvider
 import com.google.samples.apps.nowinandroid.feature.foryou.OnboardingUiState.Loading
 import com.google.samples.apps.nowinandroid.feature.foryou.OnboardingUiState.NotShown
 import com.google.samples.apps.nowinandroid.feature.foryou.OnboardingUiState.Shown
+import com.google.samples.apps.nowinandroid.roborazzi.DefaultTestDevices
+import com.google.samples.apps.nowinandroid.roborazzi.captureForDevice
+import com.google.samples.apps.nowinandroid.roborazzi.captureMultiDevice
 import dagger.hilt.android.testing.HiltTestApplication
 import org.junit.Before
 import org.junit.Rule
