+++ feature/search/build.gradle.kts
@@ -33,6 +33,7 @@ dependencies {
 
     testImplementation(projects.core.testing)
 
+    androidTestImplementation(libs.androidx.compose.ui.test)
     androidTestImplementation(projects.core.testing)
 }
