+++ feature/settings/build.gradle.kts
@@ -31,5 +31,5 @@ dependencies {
 
     testImplementation(projects.core.testing)
 
-    androidTestImplementation(projects.core.testing)
+    androidTestImplementation(libs.androidx.compose.ui.test)
 }
