+++ settings.gradle.kts
@@ -59,6 +59,7 @@ include(":feature:topic")
 include(":feature:search")
 include(":feature:settings")
 include(":lint")
+include(":roborazzi")
 include(":sync:work")
 include(":sync:sync-test")
 include(":ui-test-hilt-manifest")
