+++ sync/work/build.gradle.kts
@@ -41,5 +41,4 @@ dependencies {
     androidTestImplementation(libs.androidx.work.testing)
     androidTestImplementation(libs.hilt.android.testing)
     androidTestImplementation(libs.kotlinx.coroutines.guava)
-    androidTestImplementation(projects.core.testing)
 }
